feat: create new DevicesPage component and remove old devices page
- Implemented a new DevicesPage component with enhanced UI and functionality.
- Added device filtering and searching capabilities.
- Included device metrics display (battery level, input/output watts, temperature).
- Integrated loading and error handling states.
- Removed the old devices page to streamline the codebase.
